JB - updating manual
