+++ manual_fantasylife_jbtheshadow/data/categories.json
@@ -1,41 +1,31 @@
 {
     "$schema": "https://github.com/ManualForArchipelago/Manual/raw/main/schemas/Manual.categories.schema.json",
     "DLC": { "yaml_option": ["dlc"], "hidden": true },
-    "Single Licenses": { "yaml_option": [ "life_licenses", "!life_progressive"], "hidden": true },
-    "Fast Licenses": { "yaml_option": [ "life_licenses", "life_progressive", "life_fast"], "hidden": true },
-    "Progressive Licenses": { "yaml_option": [ "life_licenses", "life_progressive", "!life_fast" ], "hidden": true },
-    "Combat License": { "hidden": true },
-    "Easier Combat": { "hidden": true },
-    "Gathering License": { "hidden": true },
-    "Crafting License": { "hidden": true },
-    "Paladin License": { "hidden": true },
-    "Mercenary License": { "hidden": true },
-    "Hunter License": { "hidden": true },
-    "Magician License": { "hidden": true },
-    "Miner License": { "hidden": true },
-    "Woodcutter License": { "hidden": true },
-    "Angler License": { "hidden": true },
-    "Cook License": { "hidden": true },
-    "Blacksmith License": { "hidden": true },
-    "Carpenter License": { "hidden": true },
-    "Tailor License": { "hidden": true },
-    "Alchemist License": { "hidden": true },
+    "Single Licenses": { "yaml_option": [ "life_licenses", "!lives_progressive"], "hidden": true },
+    "Fast Licenses": { "yaml_option": [ "life_licenses", "lives_progressive", "lives_fast"], "hidden": true },
+    "Progressive Licenses": { "yaml_option": [ "life_licenses", "lives_progressive", "!lives_fast" ], "hidden": true },
     "Other Requests 1": { "hidden": true },
     "Other Requests 2": { "hidden": true },
     "Other Requests 3": { "hidden": true },
     "Other Requests 4": { "hidden": true },
     "Bag Bonus": { "hidden": true },
     "Storage Bonus": { "hidden": true },
-    "Shopping Bonus": { "hidden": true },
+    "Shop Bonus": { "hidden": true },
     "Pet Bonus": { "hidden": true },
-    "Mounts Bonus": { "hidden": true },
-    "Fashion Bonus": { "hidden": true },
-    "Media Bonus": { "yaml_option": ["include_happy_audio_video"], "hidden": true },
-    "Claimable Passwords": { "yaml_option": ["include_passwords"] },
-    "Progressive Bliss": { "yaml_option": ["bliss_bonuses"], "hidden": true },
+    "Animal Bonus": { "hidden": true },
+    "Customization Bonus": { "hidden": true },
+    "Theatre Bonus": { "hidden": true },
+    "Claimable Passwords": { "yaml_option": ["passwords"] },
+    "Progressive Bliss": { "yaml_option": ["bliss"], "hidden": true },
     "Level Up Checks": { "yaml_option": ["character_levels"] },
     "Skill Level Checks": { "yaml_option": ["skill_levels"] },
-    "Item Restrictions": { "yaml_option": ["enable_item_restrictions"] },
+    "Item Restrictions": { "yaml_option": ["item_restrictions"] },
+    "Life Challenges": { "yaml_option": ["life_challenges"], "hidden": true },
+    "Life Ranks": { "yaml_option": ["life_ranks"], "hidden": true },
+    "Combat": { "hidden": true },
+    "Easier Combat": { "hidden": true },
+    "Gatherer": { "hidden": true },
+    "Artisan": { "hidden": true },
     "Fledgling": { "hidden": true },
     "Apprentice": { "hidden": true },
     "Adept": { "hidden": true },
@@ -43,5 +33,17 @@
     "Master": { "hidden": true },
     "Hero": { "hidden": true },
     "Legend": { "hidden": true },
-    "Creator": { "hidden": true }
+    "Creator": { "hidden": true },
+    "Paladin": { "hidden": true },
+    "Mercenary": { "hidden": true },
+    "Hunter": { "hidden": true },
+    "Magician": { "hidden": true },
+    "Miner": { "hidden": true },
+    "Woodcutter": { "hidden": true },
+    "Angler": { "hidden": true },
+    "Cook": { "hidden": true },
+    "Blacksmith": { "hidden": true },
+    "Carpenter": { "hidden": true },
+    "Tailor": { "hidden": true },
+    "Alchemist": { "hidden": true }
 }
