+++ manual_fantasylife_jbtheshadow/hooks/Helpers.py
@@ -1,52 +1,79 @@
 from typing import Optional, Any, Union
+from unittest import case
+
 from BaseClasses import MultiWorld
 
 from .. import Helpers as rootHelpers
+from .Data import get_available_lives
 
 # Use this if you want to override the default behavior of is_option_enabled
 # Return True to enable the category, False to disable it, or None to use the default behavior
 def before_is_category_enabled(multiworld: MultiWorld, player: int, category_name: str) -> Optional[bool]:
     other_requests = rootHelpers.get_option_value(multiworld, player, "other_requests")
     goal = rootHelpers.get_option_value(multiworld, player, "goal")
-    life_max_rank = rootHelpers.get_option_value(multiworld, player, "life_max_rank")
+    lives_max_rank = rootHelpers.get_option_value(multiworld, player, "lives_max_rank")
+
+    bliss_available = rootHelpers.get_option_value(multiworld, player, "bliss_available")
+    available_lives = get_available_lives()
+
     match category_name:
         case "Other Requests 1" if other_requests < 1:
             return False
-
         case "Other Requests 2" if other_requests < 2:
             return False
-
         case "Other Requests 3" if other_requests < 3:
             return False
-
         case "Other Requests 4" if other_requests < 4:
             return False
-
         case "Wish Hunt" if goal not in (0, 2):
             return False
-
-        case "Fledgling" if life_max_rank < 1:
+        case "Fledgling" if lives_max_rank < 1:
             return False
-
-        case "Apprentice" if life_max_rank < 2:
+        case "Apprentice" if lives_max_rank < 2:
             return False
-
-        case "Adept" if life_max_rank < 3:
+        case "Adept" if lives_max_rank < 3:
             return False
-
-        case "Expert" if life_max_rank < 4:
+        case "Expert" if lives_max_rank < 4:
             return False
-
-        case "Master" if life_max_rank < 5:
+        case "Master" if lives_max_rank < 5:
             return False
-
-        case "Hero" if life_max_rank < 6:
+        case "Hero" if lives_max_rank < 6:
             return False
-
-        case "Legend" if life_max_rank < 7:
+        case "Legend" if lives_max_rank < 7:
             return False
-
-        case "Creator" if life_max_rank < 8:
+        case "Creator" if lives_max_rank < 8:
+            return False
+        case "Animal Bonus" if bliss_available == 1:
+            return False
+        case "Pet Bonus" if bliss_available == 1:
+            return False
+        case "Customization Bonus" if bliss_available == 1:
+            return False
+        case "Theatre Bonus" if bliss_available != 0:
+            return False
+        case "Paladin" if 1 not in available_lives:
+            return False
+        case "Mercenary" if 2 not in available_lives:
+            return False
+        case "Hunter" if 3 not in available_lives:
+            return False
+        case "Magician" if 4 not in available_lives:
+            return False
+        case "Miner" if 5 not in available_lives:
+            return False
+        case "Woodcutter" if 6 not in available_lives:
+            return False
+        case "Angler" if 7 not in available_lives:
+            return False
+        case "Cook" if 8 not in available_lives:
+            return False
+        case "Blacksmith" if 9 not in available_lives:
+            return False
+        case "Carpenter" if 10 not in available_lives:
+            return False
+        case "Tailor" if 11 not in available_lives:
+            return False
+        case "Alchemist" if 12 not in available_lives:
             return False
 
     return None
